requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cmake==3.18.2.post1
decorator==4.4.2
dlib==19.18.0
face-recognition==1.3.0
face-recognition-models==0.3.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
moviepy==1.0.3
numpy==1.18.4
pandas==1.3.4
opencv-python==4.4.0.46
Pillow==8.0.1
proglog==0.1.9
requests==2.24.0
tqdm==4.51.0
urllib3==1.25.11
wincertstore==0.2
